+++ lib/src/alert_dialog/show_ok_cancel_alert_dialog.dart
@@ -31,7 +31,7 @@ Future<OkCancelResult> showOkCancelAlertDialog({
   VerticalDirection actionsOverflowDirection = VerticalDirection.up,
   bool fullyCapitalizedForMaterial = true,
   bool canPop = true,
-  PopInvokedCallback? onPopInvoked,
+  PopInvokedWithResultCallback<OkCancelResult>? onPopInvokedWithResult,
   AdaptiveDialogBuilder? builder,
   RouteSettings? routeSettings,
 }) async {
@@ -55,7 +55,7 @@ Future<OkCancelResult> showOkCancelAlertDialog({
     actionsOverflowDirection: actionsOverflowDirection,
     fullyCapitalizedForMaterial: fullyCapitalizedForMaterial,
     canPop: canPop,
-    onPopInvoked: onPopInvoked,
+    onPopInvokedWithResult: onPopInvokedWithResult,
     builder: builder,
     actions: [
       AlertDialogAction(