+++ lib/src/text_input_dialog/show_text_input_dialog.dart
@@ -25,7 +25,7 @@ Future<List<String>?> showTextInputDialog({
   VerticalDirection actionsOverflowDirection = VerticalDirection.up,
   bool fullyCapitalizedForMaterial = true,
   bool canPop = true,
-  PopInvokedCallback? onPopInvoked,
+  PopInvokedWithResultCallback<List<String>?>? onPopInvokedWithResult,
   bool autoSubmit = false,
   AdaptiveDialogBuilder? builder,
   RouteSettings? routeSettings,
@@ -52,7 +52,7 @@ Future<List<String>?> showTextInputDialog({
             style: adaptiveStyle,
             useRootNavigator: useRootNavigator,
             canPop: canPop,
-            onPopInvoked: onPopInvoked,
+            onPopInvokedWithResult: onPopInvokedWithResult,
             autoSubmit: autoSubmit,
           );
           return builder == null ? dialog : builder(context, dialog);
@@ -75,7 +75,7 @@ Future<List<String>?> showTextInputDialog({
               style: adaptiveStyle,
               useRootNavigator: useRootNavigator,
               canPop: canPop,
-              onPopInvoked: onPopInvoked,
+              onPopInvokedWithResult: onPopInvokedWithResult,
               autoSubmit: autoSubmit,
             ),
           );
@@ -103,7 +103,7 @@ Future<List<String>?> showTextInputDialog({
             useRootNavigator: useRootNavigator,
             fullyCapitalized: fullyCapitalizedForMaterial,
             canPop: canPop,
-            onPopInvoked: onPopInvoked,
+            onPopInvokedWithResult: onPopInvokedWithResult,
             autoSubmit: autoSubmit,
           );
           return builder == null ? dialog : builder(context, dialog);